requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
sqlalchemy==2.0.30
psycopg2-binary==2.9.10
python-dotenv==1.0.1
pandas==2.2.2
numpy==1.26.4